Add full-stack property encryption tests and fix the metadata int/long wire mismatch
PropertyEncryptionFullStackTests drives the whole pathway through the real public API (real driver, container, and crypto; Local KES + in-memory repository; no server): round-trips for every supported property type by key alias and id, explicit/persisted AAD, wrong-AAD and tampered-ciphertext failures, fresh-IV uniqueness, unsupported types, propagate-by-type errors, and multi-profile routing.

It immediately caught a real bug the mock-based unit tests could not: EnvelopeMetadataBuilder wrote aad_protocol_major/minor as int, but PackStream decoding hands integers back as long, so EnvelopeMetadataExtractor threw MetadataExtractionException on every real decrypt. The builder now writes long, the extractor reads long, and both unit-test fixtures use the long values the wire actually produces.

+++ Neo4j.Driver/Neo4j.Driver/Internal/Encryption/EnvelopeMetadataBuilder.cs
@@ -29,8 +29,8 @@ internal class EnvelopeMetadataBuilder : IEnvelopeMetadataBuilder
             [EnvelopeMetadataKeys.KeyId] = metadata.KeyId,
             [EnvelopeMetadataKeys.Iv] = metadata.Iv,
             [EnvelopeMetadataKeys.Aad] = metadata.Aad,
-            [EnvelopeMetadataKeys.AadProtocolMajor] = metadata.AadProtocolMajor,
-            [EnvelopeMetadataKeys.AadProtocolMinor] = metadata.AadProtocolMinor
+            [EnvelopeMetadataKeys.AadProtocolMajor] = (long)metadata.AadProtocolMajor,
+            [EnvelopeMetadataKeys.AadProtocolMinor] = (long)metadata.AadProtocolMinor
         };
     }
 }
